+++ flask_site/models/__init__.py
@@ -1,0 +1,13 @@
+""" All model objects """
+# pylint: disable=import-error
+from .basic_info import BasicInfo, RankedDivisionInfo, RankedSplit, RankTier, Division
+from .ranked_game_event import RankedGameEvent
+
+__all__ = [
+    'BasicInfo',
+    'RankedGameEvent',
+    'RankedDivisionInfo',
+    'RankedSplit',
+    'RankTier',
+    'Division'
+]